feat: export BridgeDownloadMenu
