+++ cspell.json
@@ -20,6 +20,7 @@
         "EIPs",
         "Grotesk",
         "gtag",
+        "headlessui",
         "HECO",
         "imtoken",
         "keychain",
