+++ src/components/pages/Download/BridgeDownload/BridgeDownload.tsx
@@ -1,11 +1,8 @@
 import { FC, ReactNode } from 'react';
 
-import { Menu } from '@headlessui/react';
-
-import { useOneKeyVersion } from '../../../../data';
 import { Box } from '../../../base';
 
-const { Button: MenuButton } = Menu;
+import { BridgeDownloadMenu } from './BridgeDownloadMenu';
 
 export interface BridgeDownloadProps {
   children?: ReactNode;
@@ -13,35 +10,14 @@ export interface BridgeDownloadProps {
 
 export const BridgeDownload: FC<BridgeDownloadProps> = (props) => {
   const { children } = props;
-  const { formattedData } = useOneKeyVersion();
-  const { bridge } = formattedData;
-
-  const options = [
-    bridge.mac,
-    bridge.win,
-    bridge.linux32Deb,
-    bridge.linux64Deb,
-    bridge.linux32Rpm,
-    bridge.linux64Rpm,
-  ];
 
   return (
-    <div>
+    <Box>
       <Box>
-        <Menu>
-          <MenuButton<'button'>>More</MenuButton>
-          <Menu.Items<'div'>>
-            <Box>{bridge.version}</Box>
-            {options.map((option) => (
-              <Menu.Item<'div'> key={option.url}>
-                <a href={option.url}>{option.name}</a>
-              </Menu.Item>
-            ))}
-          </Menu.Items>
-        </Menu>
+        <BridgeDownloadMenu />
       </Box>
 
       {children}
-    </div>
+    </Box>
   );
 };
